+++ R/correct_attributes.R
@@ -262,61 +262,53 @@ zap_attributes.data.frame <- function(x, attributes = NULL) {
 #' Modelled on [haven::zap_labels()], zaps variable labels (not value labels).
 #'
 #' @param x the data frame or variable
+#'
 #' @export
+#' @examples
+#' x <- haven::labelled(rep(1:5, each = 1), c(Bad = 1, Good = 5))
+#' zap_label(x)
 zap_label <- function(x) {
   UseMethod("zap_label")
 }
+
+#' @export
 zap_label.data.frame <- function(x) {
   x[] <- lapply(x, zap_label)
   x
 }
+
+#' @export
 zap_label.default <- function(x) {
   attr(x, "label") <- NULL
   x
 }
 
 
-
-#' Aggregate variables and remember which variables this were
+#' Zap labelled class
 #'
-#' The resulting variables will have the attribute `scale_item_names` containing
-#' the basis for aggregation. Its `label` attribute will refer to the common stem of the
-#' aggregated variable names (if any), the number of variables, and the
-#' aggregation function.
+#' Modelled on [haven::zap_labels()], zaps labelled class (not other attributes).
 #'
-#' @param items data.frame of the items that should be aggregated
-#' @param fun aggregation function, defaults to rowMeans with na.rm = FALSE
-#' @param stem common stem for the variables, specify if it should not be auto-detected
-#' as the longest common stem of the variable names
+#' @param x the data frame or variable
 #' @export
-#' @examples
-#' testdf <- data.frame(bfi_neuro_1 = rnorm(20), bfi_neuro_2 = rnorm(20),
-#'                     bfi_neuro_3R = rnorm(20), age = rpois(20, 30))
-#' item_names <- c('bfi_neuro_1', 'bfi_neuro_2', 'bfi_neuro_3R')
-#' testdf$bfi_neuro <- aggregate_and_document_scale(testdf[, item_names])
-#' testdf$bfi_neuro
-aggregate_and_document_scale <- function(items, fun = rowMeans, stem = NULL) {
-  new_scale <- fun(items)
-  item_names <- names(items)
-  attributes(new_scale)$scale_item_names <- item_names
-
-  # find longest common stem
-  if (is.null(stem)) {
-    max_len <- min(nchar(item_names))
-    for (l in max_len:0) {
-      stem <- unique(stringr::str_sub(item_names, 1, l))
-      if (length(stem) == 1) break
-    }
-  }
-  # string trimming for idiots
-  if (nchar(stem)) {
-    stem <- stringr::str_match(stem, "^(.+?)_?$")[, 2]
-  }
-
-  attributes(new_scale)$label <- paste(ncol(items), stem, "items aggregated by",
-                                      deparse(substitute(fun)))
-  new_scale
+zap_labelled <- function(x) {
+  UseMethod("zap_labelled")
 }
+
+#' @export
+zap_labelled.data.frame <- function(x) {
+  x[] <- lapply(x, zap_labelled)
+  x
+}
+
+#' @export
+zap_labelled.labelled <- function(x) {
+  if (inherits(x, "labelled")) {
+    unclass(x)
+  } else {
+    x
+  }
+}
+
 
 
 #' Reverse labelled values
@@ -333,6 +325,14 @@ reverse_labelled_values <- function(x) {
   labels <- attributes(x)$labels
   values <- unname(labels)
   labels <- names(labels)
+  if (
+    sum(!is.na(values)) == 0 ||
+    (any(x > max(values, na.rm = TRUE) |
+          x < min(values, na.rm = TRUE), na.rm = TRUE))) {
+    warning(deparse(substitute(x)), ": There are values outside the ",
+            "labelled range. Reversion will only work if both the minimum ",
+            "and maximum of the range are part of the responses.")
+  }
   if (length(values) < length(unique(x)) ) {
     # if only some values have labels (e.g. extremes), make sure we include all
     possible_replies <- union(values, unique(x))
@@ -349,6 +349,7 @@ reverse_labelled_values <- function(x) {
         length(range)) {
       possible_replies <- range
     }
+
     possible_replies <- sort(possible_replies)
     recode_replies <- stats::setNames(
       as.list(possible_replies), rev(possible_replies))